Added new example client supplemental. This will be used to test the scenario of generating UiBinder classes in a dependency jar/gwt-lib Start of UI refactoring of examples app
