+++ gwt-uibinder-example/gwt-uibinder-example-client-supplemental/src/main/module.gwt.xml
@@ -22,23 +22,10 @@
   <inherits name="com.google.gwt.user.User" />
   <inherits name="com.google.gwt.user.theme.clean.Clean" />
 
-  <!-- to be replaced -->
-  <inherits name="com.google.gwt.uibinder.UiBinder"/>
-
   <!-- new UiBinder -->
   <inherits name="org.gwtproject.uibinder.UiBinder"/>
 
-  <inherits name="com.google.common.util.concurrent.Concurrent"/>
-  <inherits name="dagger.Dagger"/>
-  
-  <inherits name="org.gwtproject.uibinder.example.Supplemental"/>
-  
   <source path="" />
-
-  <entry-point class="org.gwtproject.uibinder.example.Application" />
-
-  <!-- Only support recent browsers -->
-  <set-property name="user.agent" value="ie10,gecko1_8,safari" />
 
   <!-- needed as GSS was hardcoded to true for UiBinder processor -->
   <set-configuration-property name="CssResource.enableGss" value="true" />
