+++ gwt-uibinder-example/gwt-uibinder-example-client/src/main/java/org/gwtproject/uibinder/example/ApplicationController.java
@@ -36,6 +36,7 @@ public class ApplicationController {
   }
 
   public void start(HasWidgets hasWidgets) {
+    hasWidgets.clear();
     hasWidgets.add(shell.asWidget());
   }
 }
