+++ gwt-uibinder-example/gwt-uibinder-example-server/src/main/webapp/index.html
@@ -4,6 +4,7 @@
   <meta charset=UTF-8>
   <title>GWT UiBinder Example</title>
 
+  <link rel="stylesheet" href="uibinderexample/css/reset.css">
   <script src="uibinderexample/uibinderexample.nocache.js"></script>
 </head>
 <body>
